refactor both data structure, so they do not depend on entities
